backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.10.5
pydantic-settings>=2.6.0
email-validator>=2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2
structlog==23.2.0
python-multipart>=0.0.20
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx>=0.28.0
upsonic>=0.60.0
aiosqlite>=0.19.0
slowapi>=0.1.9

